169 - Rename ListMain/DashboardList to StabilityTable
