+++ frontend/src/components/stabilityTable/StabilityTable.styles.js
